Chatroom V3.0
MAJOR OVERHAUL
Using resources from lab10 we restructured our project.
For now, this is the final draft, updates may be made but will only improve visual aesthetics.

Notes:
To clarify all commits are made by Jaden Ngo but the work done within the commits is equally shared between Jaden and Dennis Peng. This is due to the technological limitations of Dennis, that we have decided to code and compile everything on "Jaden" machine.